h
h

+++ Assets/Scripts/Guns/Bullets/Bullets.cs
@@ -8,7 +8,7 @@ public class Bullets : MonoBehaviour
 
     public float m_BulletForce;
     public int m_Damage;
-    public int m_UpTime;
+    public float m_UpTime;
 
     //others
